Delete unnecessary file
